Use the same defaults between tools
Summary:
The defaults for the service name and shards
are not the same. This means that you need to explicitly pass the flags for things
to work out of the box. Fixing this!

This diff makes sure that the service name in the service, is the same as the
berignei_put function that linuxfood wrote and configuration generator.

Default number of shards in service and configuration generator also match.

Reviewed By: mlgood

Differential Revision: D4203356

fbshipit-source-id: 20a8be4acc20cf4b35c49f2fa8f634ac517f87a4