Fix interactions between partial data exceptions and MISSING_TOO_MUCH_DATA
Summary:
We only want to throw exceptions on transient issues, not permanent failures.

While MISSING_TOO_MUCH_DATA correctly causes the key to be retried against a
different instance of the service, the old version of this code would eventually
throw an exception if no other instance had the relevant data.
Since MISSING_TOO_MUCH_DATA is non-recoverable error, this would cause queries
for the affected time range to fail permanently.

The new version will ultimately accept whatever is available, only throwing
exceptions if the missing data is likely to become available in the future.

Reviewed By: mlgood

Differential Revision: D4282776

fbshipit-source-id: d921c38e395e6099cbcb586fb56e4db162776863