Move bucket conversions to BucketUtils, use in BucketLogWriter
Summary: Quick second pass. All static methods were moved to BucketUtils to make the dependency tree cleaner. Looked at BucketLogWriter and replaced bucket conversions with BucketUtils. Two lines that already stagger processing are unable to be refactored, so they are to be cleaned up after/the same time as we switch on bucket staggering.

Reviewed By: swong15

Differential Revision: D4670851

fbshipit-source-id: ee175b5fa7dff19f0f62272e0730d789ef6b6152