Have BucketLogWriterTest handle unaligned buckets
Summary:
This is a part of an overall update to all the unit tests of beringei to handle unaligned buckets. I edit BucketUtils.cpp to use

  return bucket * windowSize + shardId * 60 % windowSize;

as its timestamp method (and similarly for the bucket method). Then fix broken unit tests and revert the BucketUtils change. The last diff will be to actually commit BucketUtils.cpp.

Reviewed By: scottfranklin

Differential Revision: D4872695

fbshipit-source-id: 8c8fc28e94a75b2faf4ae7c0726149c21c055be4