Switch scanning call sites from paging to subsharding
Summary:
Specifying the sharding factor without knowing how big the shard is is a
bit awkward in some cases, but it works.

Next step is to fix that either by querying Beringei for how big the shard is
or by using a feedback loop to adjust the sharding factor over time.

I won't land this until D5126728 finishes rolling out.

Reviewed By: ovoietsa

Differential Revision: D5269271

fbshipit-source-id: 68fc72da4394393cfeff66d71380a7160d511418