Make counter names more consistent
Summary:
Counters in beringei/lib currently have a '.' in front of them becuase
they assume `GorillaStatsManager` will add some prefix, while most of
those in beringei/client do not because they don't.

This removes the leading '.' from the counter names and appends it to
the prefix if one is provided.

Reviewed By: jeffdshen

Differential Revision: D5531248

fbshipit-source-id: 12ca8fcf83dbb32964d7299cb5ff91e3536882a2