Don't corrupt data if we can't read key list files (again)
Summary:
The previous attempt at this works fine if we cannot read the key list at shard
loading time (whether due to transient failures or because the file is truly
gone/empty). However, when this happens, we re-use the IDs for the "new" keys
that arrive.
If the process restarts or shard moves a second time in < 24h, we will read
the newly-generated key->id mappings and then read out data predating those
mappings that should not be associated with the created keys.

We can't trust our ability to clean up or otherwise tombstone the mismatched
data after the first restart because, after all, this situation only arises when
storage is unreliable in the first place.

Thus, this diff extends the key_list file format, adding a timestamp to each key
representing a lower bound on valid data. For very new keys, this is equal to
the first data point, while for older keys, it is set to the start of the oldest
data block. This causes the leftover data on disk to be discarded when the
second shard move happens.

This change is backward, but not forward compatible. We can read the old files,
but any new files will no longer be readable by old versions. This means the
rollout process may lose a bit of data due to shards moving from upgraded tasks
back to unupgraded tasks.

Reviewed By: jeffdshen

Differential Revision: D5732970

fbshipit-source-id: 137a1a84e661d8fc40998b1ba79b6e83b0ab7b62