correctly process the case when we fail listing completed block files
Summary:
In BucketMap::readData() we're calling findCompletedBlockFiles() which uses boost::filesystem::directory_iterator to list completed block files. On a very rare occasion (was detected only once) directory_iterator can throw, leaving the shard in a limbo state (we never switch it to OWNED).

The fix is to catch exception from findCompletedBlockFiles() and treat it as if we failed to find any files.

Reviewed By: dmaone

Differential Revision: D7453337

fbshipit-source-id: 4d2f97d7a0a2f2777cfb42d872a4495b3de6fc9a