Optimize TimeSeries::mergeValues for the same timeseries
Summary: Decompression takes a lot of time, and we can cut the number of decompressions we do in half in mergeValues by checking to see if the compressed blocks are the same before decompressing and merging. Comparing the compressed blacks should be very fast (possibly 128x faster than decompressing), so it should cost very little even when none of the blocks are the same.

Differential Revision: D7714944

fbshipit-source-id: 6c4798af9f733f70e337215f5e948485c5432e3e